requirements.txt
requests==2.27.1
pandas==1.4.2
python-telegram-bot==13.11
